security issues

+++ router/book.js
@@ -34,6 +34,7 @@ router.post("/", verify.verifytoken, async (req, res) => {
       title,
       serialNo,
       publicationDate,
+      userInfo: req.user.id,
     });
 
     const savedBook = await firstBook.save();
@@ -48,6 +49,15 @@ router.post("/", verify.verifytoken, async (req, res) => {
 
 router.delete("/:id", verify.verifytoken, async (req, res) => {
   try {
+    let book = await Book.findById(req.params.id);
+
+    if (!book) return res.status(404).json({ msg: "Book not found" });
+
+    // Making sure user post the book
+
+    if (book.userInfo.toString() !== req.user.id)
+      return res.status(401).json({ msg: "Not authorized" });
+
     const findBook = await Book.findByIdAndDelete({ _id: req.params.id });
 
     return res.status(200).json({ "Deleted Book is ": findBook });
@@ -60,6 +70,16 @@ router.delete("/:id", verify.verifytoken, async (req, res) => {
 
 router.put("/:id", verify.verifytoken, async (req, res) => {
   try {
+    let book = await Book.findById(req.params.id);
+
+    if (!book) return res.status(404).json({ msg: "Book not found" });
+
+    // Making sure user post the book
+    console.log(book.userInfo);
+    console.log(req.user.id);
+    if (book.userInfo.toString() !== req.user.id)
+      return res.status(401).json({ msg: "Not authorized" });
+
     const updateBook = await Book.findByIdAndUpdate(
       { _id: req.params.id },
       req.body
